add rating

+++ src/main/java/com/example/gadgetariumb7/dto/response/ProductCardResponse.java
@@ -20,12 +20,12 @@ public class ProductCardResponse {
     private int productPrice;
     private int discountPrice;
     private ProductStatus productStatus;
-    private Byte productRating;
+    private Double productRating;
     private int countOfReview;
     private boolean isFavorite;
     private boolean isCompared;
 
-    public ProductCardResponse(Long productId, String productImage, String productName, int count, int productPrice, ProductStatus productStatus, Byte productRating) {
+    public ProductCardResponse(Long productId, String productImage, String productName, int count, int productPrice, ProductStatus productStatus, Double productRating) {
         this.productId = productId;
         this.productImage = productImage;
         this.productName = productName;
